# String dictionary hash cache