--- a/src/common/vector_operations/vector_hash.cpp
+++ b/src/common/vector_operations/vector_hash.cpp
@@ -40,32 +40,35 @@ hash_t CombineHashScalar(hash_t a, hash_t b) {
 	return a ^ b;
 }
 
-template <bool HAS_RSEL, bool HAS_SEL_VECTOR, class T>
+template <bool HAS_RSEL, bool HAS_SEL_VECTOR, class T, bool INPUT_IS_ALREADY_HASH>
 void TightLoopHash(const T *__restrict ldata, hash_t *__restrict result_data, const SelectionVector *rsel, idx_t count,
                    const SelectionVector *__restrict sel_vector, const ValidityMask &mask) {
 	if (!mask.AllValid()) {
 		for (idx_t i = 0; i < count; i++) {
 			auto ridx = HAS_RSEL ? rsel->get_index_unsafe(i) : i;
 			auto idx = HAS_SEL_VECTOR ? sel_vector->get_index_unsafe(ridx) : ridx;
-			result_data[ridx] = HashOp::Operation(ldata[idx], !mask.RowIsValidUnsafe(idx));
+			result_data[ridx] = INPUT_IS_ALREADY_HASH ? CachedHashOp::Operation(ldata[idx])
+			                                          : HashOp::Operation(ldata[idx], !mask.RowIsValidUnsafe(idx));
 		}
 	} else {
 		for (idx_t i = 0; i < count; i++) {
 			auto ridx = HAS_RSEL ? rsel->get_index_unsafe(i) : i;
 			auto idx = HAS_SEL_VECTOR ? sel_vector->get_index_unsafe(ridx) : ridx;
-			result_data[ridx] = duckdb::Hash<T>(ldata[idx]);
+			result_data[ridx] =
+			    INPUT_IS_ALREADY_HASH ? CachedHashOp::Operation(ldata[idx]) : duckdb::Hash<T>(ldata[idx]);
 		}
 	}
 }
 
-template <bool HAS_RSEL, class T>
+template <bool HAS_RSEL, class T, bool INPUT_IS_ALREADY_HASH = false>
 void TemplatedLoopHash(Vector &input, Vector &result, const SelectionVector *rsel, idx_t count) {
 	if (input.GetVectorType() == VectorType::CONSTANT_VECTOR) {
 		result.SetVectorType(VectorType::CONSTANT_VECTOR);
 
 		auto ldata = ConstantVector::GetData<T>(input);
 		auto result_data = ConstantVector::GetData<hash_t>(result);
-		*result_data = HashOp::Operation(*ldata, ConstantVector::IsNull(input));
+		*result_data = INPUT_IS_ALREADY_HASH ? CachedHashOp::Operation(*ldata)
+		                                     : HashOp::Operation(*ldata, ConstantVector::IsNull(input));
 	} else {
 		result.SetVectorType(VectorType::FLAT_VECTOR);
 
@@ -73,13 +76,13 @@ void TemplatedLoopHash(Vector &input, Vector &result, const SelectionVector *rse
 		input.ToUnifiedFormat(count, idata);
 
 		if (idata.sel->IsSet()) {
-			TightLoopHash<HAS_RSEL, true, T>(UnifiedVectorFormat::GetData<T>(idata),
-			                                 FlatVector::GetData<hash_t>(result), rsel, count, idata.sel,
-			                                 idata.validity);
+			TightLoopHash<HAS_RSEL, true, T, INPUT_IS_ALREADY_HASH>(UnifiedVectorFormat::GetData<T>(idata),
+			                                                        FlatVector::GetData<hash_t>(result), rsel, count,
+			                                                        idata.sel, idata.validity);
 		} else {
-			TightLoopHash<HAS_RSEL, false, T>(UnifiedVectorFormat::GetData<T>(idata),
-			                                  FlatVector::GetData<hash_t>(result), rsel, count, idata.sel,
-			                                  idata.validity);
+			TightLoopHash<HAS_RSEL, false, T, INPUT_IS_ALREADY_HASH>(UnifiedVectorFormat::GetData<T>(idata),
+			                                                         FlatVector::GetData<hash_t>(result), rsel, count,
+			                                                         idata.sel, idata.validity);
 		}
 	}
 }
@@ -352,8 +355,7 @@ void TightLoopCombineHashConstant(const T *__restrict ldata, hash_t constant_has
 		for (idx_t i = 0; i < count; i++) {
 			auto ridx = HAS_RSEL ? rsel->get_index(i) : i;
 			auto idx = sel_vector->get_index(ridx);
-			auto other_hash = INPUT_IS_ALREADY_HASH ? CachedHashOp::Operation(ldata[idx])
-			                                        : HashOp::Operation(ldata[idx], !mask.RowIsValid(idx));
+			auto other_hash = INPUT_IS_ALREADY_HASH ? CachedHashOp::Operation(ldata[idx]) : duckdb::Hash<T>(ldata[idx]);
 			hash_data[ridx] = CombineHashScalar(constant_hash, other_hash);
 		}
 	}
@@ -375,8 +377,7 @@ static inline void TightLoopCombineHash(const T *__restrict ldata, hash_t *__res
 		for (idx_t i = 0; i < count; i++) {
 			auto ridx = HAS_RSEL ? rsel->get_index_unsafe(i) : i;
 			auto idx = HAS_SEL ? sel_vector->get_index_unsafe(ridx) : ridx;
-			auto other_hash = INPUT_IS_ALREADY_HASH ? CachedHashOp::Operation(ldata[idx])
-			                                        : HashOp::Operation(ldata[idx], !mask.RowIsValid(idx));
+			auto other_hash = INPUT_IS_ALREADY_HASH ? CachedHashOp::Operation(ldata[idx]) : duckdb::Hash<T>(ldata[idx]);
 			hash_data[ridx] = CombineHashScalar(hash_data[ridx], other_hash);
 		}
 	}
@@ -482,8 +483,8 @@ void CombineHashTypeSwitch(Vector &hashes, Vector &input, const SelectionVector 
 
 void VectorOperations::Hash(Vector &input, Vector &result, idx_t count) {
 	if (input.GetVectorType() == VectorType::DICTIONARY_VECTOR && DictionaryVector::CanCacheHashes(input)) {
-		DictionaryVector::GetCachedHashes(input, result);
-		result.Flatten(count);
+		VectorOperations::Copy(DictionaryVector::GetCachedHashes(input), result, DictionaryVector::SelVector(input),
+		                       count, 0, 0);
 	} else {
 		HashTypeSwitch<false>(input, result, nullptr, count);
 	}
@@ -491,9 +492,8 @@ void VectorOperations::Hash(Vector &input, Vector &result, idx_t count) {
 
 void VectorOperations::Hash(Vector &input, Vector &result, const SelectionVector &sel, idx_t count) {
 	if (input.GetVectorType() == VectorType::DICTIONARY_VECTOR && DictionaryVector::CanCacheHashes(input)) {
-		DictionaryVector::GetCachedHashes(input, result);
-		result.Slice(sel, count);
-		result.Flatten(count);
+		Vector input_hashes(DictionaryVector::GetCachedHashes(input), DictionaryVector::SelVector(input), count);
+		TemplatedLoopHash<true, hash_t, true>(input_hashes, result, &sel, count);
 	} else {
 		HashTypeSwitch<true>(input, result, &sel, count);
 	}
@@ -501,8 +501,7 @@ void VectorOperations::Hash(Vector &input, Vector &result, const SelectionVector
 
 void VectorOperations::CombineHash(Vector &hashes, Vector &input, idx_t count) {
 	if (input.GetVectorType() == VectorType::DICTIONARY_VECTOR && DictionaryVector::CanCacheHashes(input)) {
-		Vector input_hashes(LogicalType::HASH, false, false);
-		DictionaryVector::GetCachedHashes(input, input_hashes);
+		Vector input_hashes(DictionaryVector::GetCachedHashes(input), DictionaryVector::SelVector(input), count);
 		TemplatedLoopCombineHash<false, hash_t, true>(input_hashes, hashes, nullptr, count);
 	} else {
 		CombineHashTypeSwitch<false>(hashes, input, nullptr, count);
@@ -511,8 +510,7 @@ void VectorOperations::CombineHash(Vector &hashes, Vector &input, idx_t count) {
 
 void VectorOperations::CombineHash(Vector &hashes, Vector &input, const SelectionVector &rsel, idx_t count) {
 	if (input.GetVectorType() == VectorType::DICTIONARY_VECTOR && DictionaryVector::CanCacheHashes(input)) {
-		Vector input_hashes(LogicalType::HASH, false, false);
-		DictionaryVector::GetCachedHashes(input, input_hashes);
+		Vector input_hashes(DictionaryVector::GetCachedHashes(input), DictionaryVector::SelVector(input), count);
 		TemplatedLoopCombineHash<true, hash_t, true>(input_hashes, hashes, &rsel, count);
 	} else {
 		CombineHashTypeSwitch<true>(hashes, input, &rsel, count);