This PR implements caching for hashes of string dictionaries. This avoid hashing the same values twice, which can be expensive, especially for long strings.

## Implementation
When creating a dictionary `Vector` from a storage dictionary using `Vector::Dictionary`, we now create an uninitialized `VectorChildBuffer` to store the hashes. If `VectorOperations::Hash` is called on the dictionary `Vector`, we check if this buffer is there, and if it is initialized. If it's uninitialized, we hash the storage dictionary, and cache the hashes in the buffer. If it's initialized, we simply grab the cached hashes.

This essentially gets us half the benefit of the (now closed) PR #18184 for string dictionaries that are used as keys to aggregation (as shown in the benchmark below). However, this PR has a tiny code footprint, and also applies to dictionaries with higher unique counts (for example, it speeds up Q34 of ClickBench by ~5%).

## Benchmark
```sql
-- generate similar data to PR #18184
create or replace table test as
select
  format('{:a>256}', cast(range % 100 as varchar)) a256, a256 b256,
  a256[-128:] a128, a128 b128,
  a256[-64:] a64, a64 b64,
  a256[-32:] a32, a32 b32,
  a256[-16:] a16, a16 b16
from range(10_000_000);
-- run each query 3 times
.mode trash
.timer on
select a16, b16 from test group by all;select a16, b16 from test group by all;select a16, b16 from test group by all;
select a32, b32 from test group by all;select a32, b32 from test group by all;select a32, b32 from test group by all;
select a64, b64 from test group by all;select a64, b64 from test group by all;select a64, b64 from test group by all;
select a128, b128 from test group by all;select a128, b128 from test group by all;select a128, b128 from test group by all;
select a256, b256 from test group by all;select a256, b256 from test group by all;select a256, b256 from test group by all;
```

Average of 3 runs:

| String Length | Current [s] | This PR [s] | Speedup [x] |
|-:|-:|-:|-:|
| 16 | 0.031 | **0.024** | 1.29x |
| 32 | 0.028 | **0.020** | 1.40x |
| 64 | 0.034 | **0.024** | 1.42x |
| 128 | 0.051 | **0.028** | 1.82x |
| 256 | 0.088 | **0.039** | 2.26x |

I have no idea why the strings of length 32 are faster (for both Current and This PR).

Anyways, we get a 2.26x speedup for the longest strings. The benefit comes from caching the hashes. PR #18184 has a >5x speedup, as it also benefits from using pointer comparisons instead of full string comparisons.